Instrum.py runable from command line.
I just added a main method (I think a little better style) and call it.

I scrapped the idea of having an arugment for output file. It just takes
the code to be instrumented as the only input.

Really weird problem with the #!/usr/bin/python at the top.. eniac won't
execute it and I don't know why. I suspect something with Unix/Windows
newline difference..